gallery menu upgrade

+++ src/main/java/tgside/handlers/TGTextHandler.java
@@ -21,30 +21,31 @@ public class TGTextHandler extends TGHandler {
     @Override
     public void handlIt() {
         String text = msg.getText();
-        switch (text) {
-            case "/start":
-                sendMessage("Just drop a new photo of a growlamp");
-                break;
-            case "score":
-                sendMessage("score-table generator is not ready yet :(");
-                break;
-            case "delete last":
-                new TGPhotoDel(msg, bot).handlIt();
-                break;
-            case "\uD83D\uDD34 delete":
-                boolean wasDelet = false;
-                try {
-                    wasDelet = new TGPhotoDel(msg, bot).deleteLastPhoto();
-                    sendMessage("Photo has been deleted");
-                } catch (PhotoListIsEmptyException e) {
-                    sendMessage("delete error %(. photoalbum is empty");
-                    e.printStackTrace();
-                }
-                break;
-            case "gallery":
-                new TGGallery(msg, bot).handlIt();
-            default:
-                sendMessage("Just drop a new photo of a growlamp");
+        if (text.equals("/start")) {
+            sendMessage("Just drop a new photo of a growlamp");
+
+        } else if (text.equals("score")) {
+            sendMessage("score-table generator is not ready yet :(");
+
+        } else if (text.equals("delete last")) {
+            new TGPhotoDel(msg, bot).handlIt();
+
+        } else if (text.equals("\uD83D\uDD34 delete")) {
+            boolean wasDelet = false;
+            try {
+                wasDelet = new TGPhotoDel(msg, bot).deleteLastPhoto();
+                sendMessage("Photo has been deleted");
+            } catch (PhotoListIsEmptyException e) {
+                sendMessage("delete error %(. photoalbum is empty");
+                e.printStackTrace();
+            }
+
+        } else if (text.equals("gallery")) {
+            new TGGallery(msg, bot).show();
+        } else if (text.matches("^@.*")) {
+            new TGGallery(msg, bot).handlIt(text.substring(1));
+        } else {
+            sendMessage("Just drop a new photo of a growlamp");
         }
     }
 }
